Revert "feat: add report and decay pages (#67)"
This reverts commit 9742eefa9fe93f897dceb2f1f3f4bc313bdefac9.
